backend/requirements.txt
# ================================
# 🌐 WEB FRAMEWORK & API
# ================================
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# ================================  
# 🗄️ DATABASE & ORM (ASYNC)
# ================================
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.4
psycopg2-binary==2.9.10

# ================================
# 🚀 CACHE & SESSION STORE
# ================================
redis==6.4.0

# ================================
# 🔐 AUTHENTICATION & SECURITY  
# ================================
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
cryptography>=45.0.6

# Rate limiting
slowapi==0.1.9

# ================================
# 🤖 AI & AGENTS
# ================================
# Microsoft Agent Framework (next-gen successor to AutoGen)
# Note: Using beta versions as stable releases are not yet available
agent-framework>=1.0.0rc1
agent-framework-core>=1.0.0rc1
# agent-framework-azure-ai>=0.1.0  # Optional: uncomment if using Azure AI features

# AI Model Providers
openai>=1.100.1
anthropic==0.64.0
tiktoken>=0.11.0

# Legacy AutoGen (kept temporarily for compatibility layer during migration)
# TODO: Remove after full migration is complete
autogen-agentchat==0.7.3
autogen-core==0.7.3
autogen-ext==0.7.3

# ================================
# 🔍 VECTOR SEARCH
# ================================
pgvector==0.4.1
numpy==2.3.2
sentence-transformers==5.1.0

# ================================
# ⏰ SCHEDULING & BACKGROUND TASKS
# ================================
apscheduler==3.11.0

# ================================
# 🌐 HTTP & NETWORKING
# ================================
httpx>=0.27.0,<1.0.0
resend>=2.5.0
aiofiles==24.1.0
websockets==15.0.1
stripe==10.12.0

# ================================
# 🛠️ UTILITIES & HELPERS
# ================================
python-dotenv==1.1.1
python-multipart==0.0.20
email-validator==2.2.0
pytz==2025.2
tenacity==9.1.2
watchdog==6.0.0
pyyaml==6.0.2
nest-asyncio==1.6.0

# ================================
# 🧪 DEVELOPMENT & TESTING
# ================================
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
playwright==1.54.0

# ================================
# 🚀 PERFORMANCE
# ================================
orjson==3.11.2
psutil==7.0.0

# ================================
# 🔒 SECURITY
# ================================
# slowapi already listed under Rate Limiting

# ================================
# 📊 OBSERVABILITY & MONITORING
# ================================
prometheus-client>=0.22.1
structlog>=25.4.0
opentelemetry-api>=1.34.1
opentelemetry-sdk>=1.34.1
sentry-sdk[fastapi]>=2.13.0
opentelemetry-instrumentation-fastapi>=0.55b1
opentelemetry-exporter-otlp>=1.34.1

# ================================
# 🛠️ DEV TOOLS
# ================================
ruff>=0.11.0
